backend: 补充了汇编文本生成的注释
